qr code cleanup
