+++ app/src/main/java/com/example/hive/Models/QRCode.java
@@ -7,6 +7,11 @@ import android.util.Log;
 import com.google.firebase.Firebase;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.SetOptions;
+import com.google.zxing.BarcodeFormat;
+import com.google.zxing.MultiFormatWriter;
+import com.google.zxing.WriterException;
+import com.google.zxing.common.BitMatrix;
+import com.journeyapps.barcodescanner.BarcodeEncoder;
 
 import java.io.ByteArrayOutputStream;
 import java.util.HashMap;
@@ -17,6 +22,42 @@ import java.util.HashMap;
  * store it in the db, and make it retrievable from the db
  */
 public class QRCode {
+    private String firebaseID;
+
+    /**
+     * constructor
+     * @param firebaseID
+     */
+    public QRCode(String firebaseID) {
+        this.firebaseID = firebaseID;
+    }
+
+    public String getFirebaseID() {
+        return firebaseID;
+    }
+
+    public void setFirebaseID(String firebaseID) {
+        this.firebaseID = firebaseID;
+    }
+
+    /**
+     * Generates a QR code image from event id that is automatically generated by firestore.
+     *
+     * @param width  The width of the QR code.
+     * @param height The height of the QR code.
+     * @return The generated QR code as a Bitmap.
+     * @throws WriterException If an error occurs during QR code generation.
+     */
+    public Bitmap generateQRCode(int width, int height) throws WriterException {
+        if (firebaseID == null || firebaseID.isEmpty()) {
+            throw new IllegalStateException("firebaseID is not set. Unable to generate QR code.");
+        }
+        String data = firebaseID;
+        MultiFormatWriter writer = new MultiFormatWriter();
+        BitMatrix matrix = writer.encode(data, BarcodeFormat.QR_CODE, width, height);
+        BarcodeEncoder encoder = new BarcodeEncoder();
+        return encoder.createBitmap(matrix);
+    }
 
     /**
      * Convert QR code bitmap to string
