+++ gradle/libs.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-agp = "8.7.1"
+agp = "8.6.0"
 core = "3.5.2"
 firebaseBomVersion = "33.5.1"
 compiler = "4.11.0"
